Create a VHSLocal for test utilities
